Renamed FPSCounter to FPScounter (casing).
