+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 .DS_store
+.DS_Store
 coverage
