+++ src/lib/setup/base.js
@@ -23,7 +23,7 @@ import Circle from '../../components/Circle.js'
 import RouterView from '../../components/RouterView.js'
 import Sprite from '../../components/Sprite.js'
 import Text from '../../components/Text.js'
-import FPSCounter from '../../components/FPSCounter.js'
+import FPScounter from '../../components/FPScounter.js'
 import eventListeners from '../eventListeners.js'
 import { default as log, Log } from '../log.js'
 import symbols from '../symbols.js'
@@ -136,7 +136,7 @@ export default (component) => {
         RouterView: RouterView(),
         Sprite: Sprite(),
         Text: Text(),
-        FPSCounter: FPSCounter(),
+        FPScounter: FPScounter(),
       },
       writable: false,
       enumerable: false,
